refactor: extract root view from main and changed names of views
